+++ tests/conftest.py
@@ -13,8 +13,9 @@ import pytest
 @pytest.fixture(autouse=True)
 def auto_enable_custom_integrations(enable_custom_integrations):
     """PHACC requires this fixture to be active for HA to find the
-    integration under custom_components/."""
-    yield
+    integration under custom_components/.
+    """
+    return
 
 
 @pytest.fixture(autouse=True)